requirements.txt
python-Levenshtein
regex
anytree
dockerfile-parse
pathlib
Matching
GitPython
pandas
tqdm
GitPython
flask
werkzeug
Flask-SQLAlchemy
Flask-Migrate